requirements.txt
beautifulsoup4==4.12.3
camelot-py==1.0.9
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
et_xmlfile==2.0.0
ghostscript==0.8.1
lxml==6.0.2
numpy==2.2.6
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
pandas==2.3.3
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
pycparser==2.23
pypdf==5.9.0
PyPDF2==3.0.1
pypdfium2==5.1.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
six==1.17.0
tabulate==0.9.0
typing_extensions==4.15.0
tzdata==2025.2
